+++ czkawka_core/src/common.rs
@@ -503,6 +503,16 @@ pub fn prepare_thread_handler_common(
     (progress_thread_sender, progress_thread_run, atomic_counter, check_was_stopped)
 }
 
+#[inline]
+pub fn check_if_stop_received(stop_receiver: Option<&crossbeam_channel::Receiver<()>>) -> bool {
+    if let Some(stop_receiver) = stop_receiver {
+        if stop_receiver.try_recv().is_ok() {
+            return true;
+        }
+    }
+    false
+}
+
 #[fun_time(message = "send_info_and_wait_for_ending_all_threads", level = "debug")]
 pub fn send_info_and_wait_for_ending_all_threads(progress_thread_run: &Arc<AtomicBool>, progress_thread_handle: JoinHandle<()>) {
     progress_thread_run.store(false, Ordering::Relaxed);